01_data/dtpdb/requirements.txt
Pillow==7.2.0
psycopg2-binary==2.8.6
python-dotenv==0.14.0